// src/app/models/simulation.model.ts
export interface SimulationResult {
  flexBenefitsPercentage: number;
  salaryBase: number;
  IHT: number;
  duodecimoSF: number;
  duodecimoSN: number;
  irsSF: number;
  irsSN: number;
  irs: number;
  netSalary: number;
  monthlyValueToBenefits: number;
  monthlyMealAllowance: number;
  totalMax: number;
  totalMin: number;
  salaryBaseAndIHT: number;
  rendimento: number;
  custoAnualParaEmpresa: number;
}